+++ examples/cypress/src/gen-v4/models.ts
@@ -103,7 +103,7 @@ export type Customer = {
   /**
    * @type array | undefined
    */
-  address?: Address[]
+  address?: Array<Address>
 }
 
 /**
@@ -212,11 +212,11 @@ export type Pet = {
   /**
    * @type array
    */
-  photoUrls: string[]
+  photoUrls: Array<string>
   /**
    * @type array | undefined
    */
-  tags?: Tag[]
+  tags?: Array<Tag>
   /**
    * @description pet status in the store
    * @type string | undefined
@@ -253,11 +253,11 @@ export type AddPetRequest = {
   /**
    * @type array
    */
-  photoUrls: string[]
+  photoUrls: Array<string>
   /**
    * @type array | undefined
    */
-  tags?: Tag[]
+  tags?: Array<Tag>
   /**
    * @description pet status in the store
    * @type string | undefined
@@ -300,7 +300,7 @@ export type PetNotFound = {
 /**
  * @type array
  */
-export type UserArray = User[]
+export type UserArray = Array<User>
 
 /**
  * @type object
@@ -440,7 +440,7 @@ export type AddPetResponse = AddPetStatus200 | AddPetStatus405
 /**
  * @type array
  */
-export type OptionsFindPetsByStatusStatus200 = Pet[]
+export type OptionsFindPetsByStatusStatus200 = Array<Pet>
 
 /**
  * @type object
@@ -478,7 +478,7 @@ export type FindPetsByStatusQueryStatus = ('available' | 'pending' | 'sold') | u
 /**
  * @type array
  */
-export type FindPetsByStatusStatus200 = Pet[]
+export type FindPetsByStatusStatus200 = Array<Pet>
 
 /**
  * @type any
@@ -521,7 +521,7 @@ export type FindPetsByStatusResponse = FindPetsByStatusStatus200 | FindPetsBySta
  * @description Tags to filter by
  * @type array | undefined
  */
-export type FindPetsByTagsQueryTags = string[] | undefined
+export type FindPetsByTagsQueryTags = Array<string> | undefined
 
 /**
  * @description to request with required page number or pagination
@@ -538,7 +538,7 @@ export type FindPetsByTagsQueryPageSize = string | undefined
 /**
  * @type array
  */
-export type FindPetsByTagsStatus200 = Pet[]
+export type FindPetsByTagsStatus200 = Array<Pet>
 
 /**
  * @type any
@@ -1110,7 +1110,7 @@ export type CreateUsersWithListInputStatusDefault = any
 /**
  * @type array | undefined
  */
-export type CreateUsersWithListInputData = User[] | undefined
+export type CreateUsersWithListInputData = Array<User> | undefined
 
 /**
  * @type object