+++ examples/typescript/src/gen10/models/Customer.ts
@@ -37,5 +37,5 @@ export type Customer = {
   /**
    * @type array | undefined
    */
-  address?: Address[]
+  address?: Array<Address>
 }