+++ examples/typescript/src/gen6/ts/models/Customer.ts
@@ -45,5 +45,5 @@ export type Customer = {
   /**
    * @type array | undefined
    */
-  address?: Address[]
+  address?: Array<Address>
 }